Separeted functions in more significant file names - Signals handling working with global variable code store - Some header fixes to make ir shorter and more readable - Pend 'split' input in tokens (token control working in all cases) - Pend include builtins and check functioning - Pend expansions, redirections, pipes and heredocs - Pend environment variables and export - Pend history and history expansion
